requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.6
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
Markdown==3.3.3
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
